Mental Disorders
If Neurotoxin food can ruin one's life, why do we consume so much of these toxic food? Are we obtuse or ignorant to the danger of these food? Or, is there an obscure secret that the government is withholding from the common masses? Are those in power clandestine in meetings to arcanum the truth about how much toxicity is produced in our food that our government gives a seal of endorsement?